Split up channels component
