+++ src/pages/Index.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { useAppSelector } from "../app/hooks";
 import ConnectionState from "../components/ConnectionState";
 import Loading from "../components/Loading";
-import Channels from "../features/channels/Channels";
-import Servers from "../features/servers/Servers";
+import ChannelList from "../features/channels/ChannelList";
+import ServerList from "../features/servers/ServerList";
 import { selectUser } from "../features/user/userSlice";
 
 export default function Index(): JSX.Element {
@@ -15,8 +15,8 @@ export default function Index(): JSX.Element {
                 <Loading />
             ) : (
                 <div className="flex max-h-screen w-full">
-                    <Servers />
-                    <Channels />
+                    <ServerList />
+                    <ChannelList />
                 </div>
             )}
             <ConnectionState />
